LayoutTests/imported/w3c/web-platform-tests/mathml/presentation-markup/mpadded/mpadded-003-expected.txt


PASS lspace/voffset shifts don't affect mpadded preferred width
PASS lspace/voffset shifts don't affect mpadded size
PASS content is shifted by the specified lspace/voffset
PASS content is shifted by the specified lspace/voffset (RTL)
FAIL dynamic changes of lspace/voffset assert_approx_equals: attach lspace expected 5 +/- 1 but got 0
